shop/requirements.txt
Django==3.2.9
Pillow==8.4.0
asgiref==3.3.1
django-crispy-forms==1.13.0
pip==21.3.1
pytz==2020.4
setuptools==58.4.0
sqlparse==0.4.1